# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ProposalForge & AgentLead API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Log in
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  # ─── PF DASHBOARD ────────────────────────────────────────────────────────────
...
  /pf/clients:
    get:
      operationId: listPfClients
      tags: [pf-clients]
      summary: List clients
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: search
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PfClient"
    post:
      operationId: createPfClient
      tags: [pf-clients]
      summary: Create client
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePfClientBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PfClient"
  /pf/clients/{id}:
    get:
      operationId: getPfClient
      tags: [pf-clients]
      summary: Get client
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PfClient"
    put:
      operationId: updatePfClient
      tags: [pf-clients]
      summary: Update client
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePfClientBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PfClient"
    delete:
      operationId: deletePfClient
      tags: [pf-clients]
      summary: Delete client
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  # ─── PF PROPOSALS ─────────────────────────────────────────────────────────────
...
  /pf/contracts:
    get:
      operationId: listPfContracts
      tags: [pf-contracts]
      summary: List contracts
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [draft, sent, signed, cancelled]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PfContract"
    post:
      operationId: createPfContract
      tags: [pf-contracts]
      summary: Create contract
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePfContractBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PfContract"
  /pf/contracts/{id}:
    get:
      operationId: getPfContract
      tags: [pf-contracts]
      summary: Get contract
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PfContract"
    put:
      operationId: updatePfContract
      tags: [pf-contracts]
      summary: Update contract
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePfContractBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PfContract"
    delete:
      operationId: deletePfContract
      tags: [pf-contracts]
      summary: Delete contract
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /pf/contracts/{id}/status:
    patch:
      operationId: updatePfContractStatus
      tags: [pf-contracts]
      summary: Update contract status
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStatusBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PfContract"

  # ─── PF DEPOSITS ──────────────────────────────────────────────────────────────
...
  /al/reminders:
    get:
      operationId: listAlReminders
      tags: [al-reminders]
      summary: List reminders
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: done
          schema:
            type: boolean
        - in: query
          name: upcoming
          schema:
            type: boolean
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AlReminder"
    post:
      operationId: createAlReminder
      tags: [al-reminders]
      summary: Create reminder
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAlReminderBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlReminder"
  /al/reminders/{id}:
    put:
      operationId: updateAlReminder
      tags: [al-reminders]
      summary: Update reminder
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAlReminderBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlReminder"
    delete:
      operationId: deleteAlReminder
      tags: [al-reminders]
      summary: Delete reminder
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /al/reminders/{id}/done:
    patch:
      operationId: markAlReminderDone
      tags: [al-reminders]
      summary: Mark reminder as done
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlReminder"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    RegisterBody:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        name:
          type: string
      required: [email, password, name]
    LoginBody:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required: [email, password]
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required: [token, user]
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        name:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, email, name, createdAt]
    UpdateStatusBody:
      type: object
      properties:
        status:
          type: string
      required: [status]
...
    PfClient:
      type: object
      properties:
        id:
          type: string
        companyName:
          type: string
        contactName:
          type: string
        email:
          type: string
        phone:
          type: string
        address:
          type: string
        notes:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, companyName, contactName, email, createdAt, updatedAt]
    CreatePfClientBody:
      type: object
      properties:
        companyName:
          type: string
        contactName:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        address:
          type: string
        notes:
          type: string
      required: [companyName, contactName, email]
...
    PfContract:
      type: object
      properties:
        id:
          type: string
        proposalId:
          type: string
        clientId:
          type: string
        clientName:
          type: string
        title:
          type: string
        status:
          type: string
          enum: [draft, sent, signed, cancelled]
        content:
          type: string
        signedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, clientId, clientName, title, status, content, createdAt, updatedAt]
    CreatePfContractBody:
      type: object
      properties:
        proposalId:
          type: string
        clientId:
          type: string
        title:
          type: string
        content:
          type: string
      required: [clientId, title, content]
...
    AlReminder:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        dueAt:
          type: string
          format: date-time
        done:
          type: boolean
        leadId:
          type: string
        dealId:
          type: string
        propertyId:
          type: string
        notes:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, title, dueAt, done, createdAt]
    CreateAlReminderBody:
      type: object
      properties:
        title:
          type: string
        dueAt:
          type: string
          format: date-time
        leadId:
          type: string
        dealId:
          type: string
        propertyId:
          type: string
        notes:
          type: string
      required: [title, dueAt]